webui/requirements.txt
docker
Django==3.2.9
httpx
djangorestframework
